refactor(performance): reduce function calls
The number of function calls during sampling has been reduced by merging 
the code of py_code into py_frame. This prevented further calls to 
malloc/free. The code from py_code has also been annotaded for inlining 
to further reduce the number of calls in the stack.

The copy_memory function, which gets called very frequently, has also 
been moved into the header file so that it can be inlined as well in all 
the translation units that make use of it.
