+++ src/mem.h
@@ -26,6 +26,27 @@
 
 #include <sys/types.h>
 
+#include "platform.h"
+
+#if defined(PL_LINUX)
+  #define _GNU_SOURCE
+  #include <sys/uio.h>
+  ssize_t process_vm_readv(
+    pid_t, const struct iovec *, unsigned long liovcnt,
+    const struct iovec *remote_iov, unsigned long riovcnt, unsigned long flags
+  );
+
+#elif defined(PL_WIN)
+  #include <windows.h>
+
+#elif defined(PL_MACOS)
+  #include <mach/mach.h>
+  #include <mach/mach_vm.h>
+  #include <mach/machine/kern_return.h>
+
+#endif
+
+#include "logging.h"
 
 #define OUT_OF_BOUND                  -1
 
@@ -74,7 +95,46 @@ typedef struct {
  * @param void *  the destination buffer, expected to be at least as large as
  *                the number of bytes to read.
  */
-ssize_t
-copy_memory(pid_t, void *, ssize_t, void *);
+static inline ssize_t
+copy_memory(pid_t pid, void * addr, ssize_t len, void * buf) {
+  #if defined(PL_LINUX)                                              /* LINUX */
+  struct iovec local[1];
+  struct iovec remote[1];
+
+  local[0].iov_base = buf;
+  local[0].iov_len = len;
+  remote[0].iov_base = addr;
+  remote[0].iov_len = len;
+
+  return process_vm_readv(pid, local, 1, remote, 1, 0);
+
+  #elif defined(PL_WIN)                                                /* WIN */
+  size_t n;
+  int ret = ReadProcessMemory((HANDLE) pid, addr, buf, len, &n) ? n : -1;
+  return ret;
+
+  #elif defined(PL_MACOS)                                              /* MAC */
+  mach_port_t task;
+  if (task_for_pid(mach_task_self(), pid, &task) != KERN_SUCCESS) {
+    log_d(
+      "Failed to obtain task from PID. Are you running Austin with the right "
+      "privileges?"
+    );
+    return -1;
+  }
+
+  mach_vm_size_t nread;
+  kern_return_t kr = mach_vm_read_overwrite(
+    task, (mach_vm_address_t) addr, len, (mach_vm_address_t) buf, &nread
+  );
+  if (kr != KERN_SUCCESS) {
+    log_t("copy_memory: mach_vm_read_overwrite returned %d", kr);
+    return -1;
+  }
+
+  return nread;
+
+  #endif
+}
 
 #endif // MEM_H
